+++ Bow.h
@@ -16,7 +16,7 @@ private:
 
 public:
     Bow();
-    Bow(int damage, int fireRate, int projSize, int projSpeed, int range);
+    Bow(int damage, float fireRate, int projSize, int projSpeed, int range);
     sf::Vector2f calculateRange(sf::Vector2f position, sf::Vector2f destination);
     std::optional<Projectile> attack(sf::Vector2f position, sf::Vector2f destination, bool isPlayer) override;
 };
